Implement Opera/OperaGX support and refactor DLL-injector architecture
- Refactored `payload.dll` into a minimal v20 (App-Bound) key extraction payload.
- Ported SQLite extraction and DPAPI (v10) decryption logic to `injector.exe`.
- Added support for Opera and Opera GX browsers with updated path discovery.
- Fixed a bug where v10 data was incorrectly truncated by 32 bytes.
- Improved profile discovery to handle root-level databases in Opera.
- Updated README with new build instructions and linked libraries.
